requirements.txt
matplotlib
pyyaml
requests
schedule
discord_webhook